requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LangGraph and LangChain
langgraph==0.0.62
langchain-core==0.1.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
numpy==1.24.3

# OpenAI
openai==1.3.5

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
